added functions to deal with loggedIn state
when moving from page to page the developer need to save the page's name so when refreshing the page, the iframe will get back to the page he was.

added login option -
if user not exist, you can see it.
if user exist, it will login and save the state of loggedIn.