added changes in fronEnd
changed the main script location - it was a t the end of the body and now its next to the style element.
added a button listener for the enter button in the loginForm.
signInBtn now wont be available after clicking it once, it will be available again if there was an error with the login or refreshing the page.